Refactoring permission service

+++ internal/app/v1app/permission/service/delete.go
@@ -1,7 +1,16 @@
 package service
 
-import "context"
+import (
+	"context"
 
+	"github.com/eliofery/golang-grpc/internal/core/server/grpc/interceptor"
+)
+
+// Delete ...
 func (s *service) Delete(ctx context.Context, id int64) error {
+	if !interceptor.IsAccess(ctx, deletePermission) {
+		return interceptor.ErrAccessDenied
+	}
+
 	return s.permissionRepository.Delete(ctx, id)
 }
